Test Unitaire package Utilitaire v2

+++ WorkspaceJava/tutorial/src/main/java/testUtilitaire/ConversionImpl.java
@@ -2,13 +2,13 @@ package testUtilitaire;
 
 public class ConversionImpl {
 	
-	public static double getForceFromUV(double u, double v){
-		return Math.sqrt(Math.pow(u, 2)+Math.pow(v, 2));
+	public static float getForceFromUV(float u, float v){
+		return (float) Math.sqrt(Math.pow(u, 2)+Math.pow(v, 2));
 		
 	}
 	
-	public static double getAngleFromUV(double u, double v){
-		return Math.toDegrees(Math.atan2(v, u));
+	public static float getAngleFromUV(float u, float v){
+		return (float) Math.toDegrees(Math.atan2(v, u));
 		
 	}
